+++ packages/getbloc_test/lib/getbloc_test.dart
@@ -2,4 +2,5 @@ library getbloc_test;
 
 export 'package:mocktail/mocktail.dart';
 
+export 'src/mock_controller.dart';
 export 'src/test_controller.dart';