+++ packages/getbloc_test/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.2.0] - 2021-07-25
+
+* Update getbloc 1.2.0.
+* Add Mock Controller.
+* Add MockInternalFinalCallback Getx lifecycle.
+* Add whenState to create a stub controller's state.
+
 ## [1.0.1] - 2021-06-13
 
 * Update getbloc 1.1.0.